# glooctl: Improve performance of the check command by optimizing the creation of clients 