--- a/changelog/v1.18.0-beta2/optimize-glooctl-check.yaml
+++ b/changelog/v1.18.0-beta2/optimize-glooctl-check.yaml
@@ -1,0 +1,5 @@
+changelog:
+- type: FIX
+  issueLink: https://github.com/solo-io/gloo/issues/9673
+  resolvesIssue: true
+  description: Optimizes the `glooctl check` command by reducing the time taken to check resources by almost half in large environments consisting of over 500 namespaces