# Description

Optimize `glooctl check` by creating the resource client prior to iterating over namespaces.
This saves a considerable amount of time in large envs consisting of over 700 ns and cuts time by ~50%

# Context

See slack conversation [here](https://solo-io-corp.slack.com/archives/CEDCS8TAP/p1718935489580349)

## Testing steps
In an env of 700+ ns

glooctl version: `1.17.0-beta34(latest version)`
```
glooctl check  251.22s user 266.20s system 37% cpu 22:48.24 total
```
glooctl version: `1.0.0-ci1 (This PR)`
```
glooctl-darwin-arm64-david check  81.22s user 74.02s system 24% cpu 10:28.20 total
```

# Checklist:

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works

<!---
# Author reminders (delete before opening)
- Include a concise, user-facing changelog (for details, see https://github.com/solo-io/go-utils/tree/main/changelogutils) referencing the issue that is resolved
  - Include `resolvesIssue: false` unless the issue does not require a release to be resolved; only a subset of non-user-facing issues can be considered resolved without release 
- Run codegen via `make -B install-go-tools generated-code`
- Follow guidelines laid out in the Gloo Edge [contribution guide](https://docs.solo.io/gloo-edge/latest/contributing/)
- If not ready for review, open a draft PR or apply the `work in progress` label
-->
BOT NOTES: 
resolves https://github.com/solo-io/gloo/issues/9673